Route the tui through the main binary
